AddChildPage: split up into components
